+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from telegram_upload.utils import scantree, sizeof_fmt
+from tgupi.utils import scantree, sizeof_fmt
 
 
 class TestSizeOfFmt(unittest.TestCase):
@@ -16,18 +16,18 @@ class TestSizeOfFmt(unittest.TestCase):
 
 
 class TestScanTree(unittest.TestCase):
-    @patch('telegram_upload.utils.scandir', return_value=[])
+    @patch('tgupi.utils.scandir', return_value=[])
     def test_empty_directory(self, m):
         self.assertEqual(list(scantree('foo')), [])
 
-    @patch('telegram_upload.utils.scandir')
+    @patch('tgupi.utils.scandir')
     def test_files(self, m):
         file = Mock()
         file.is_dir.return_value = False
         m.return_value = [file] * 3
         self.assertEqual(list(scantree('foo')), m.return_value)
 
-    @patch('telegram_upload.utils.scandir')
+    @patch('tgupi.utils.scandir')
     def test_directory(self, m):
         directory = Mock()
         directory.is_dir.side_effect = [True, False]