+++ tgupi/config.py
@@ -3,9 +3,9 @@ import os
 
 import click
 
-CONFIG_DIRECTORY = os.environ.get('TELEGRAM_UPLOAD_CONFIG_DIRECTORY', '~/.config')
-CONFIG_FILE = os.path.expanduser(f'{CONFIG_DIRECTORY}/telegram-upload.json')
-SESSION_FILE = os.path.expanduser(f'{CONFIG_DIRECTORY}/telegram-upload')
+CONFIG_DIRECTORY = os.environ.get('TGUPI_CONFIG_DIRECTORY', '~/.config')
+CONFIG_FILE = os.path.expanduser(f'{CONFIG_DIRECTORY}/tgupi.json')
+SESSION_FILE = os.path.expanduser(f'{CONFIG_DIRECTORY}/tgupi')
 
 
 def prompt_config(config_file):