+++ tgupi/upload_files.py
@@ -7,17 +7,17 @@ from typing import TYPE_CHECKING
 
 import click
 
-from telegram_upload.caption_formatter import CaptionFormatter, FilePath
-from telegram_upload.constants import SPLIT_FILE_PART_NUMBER_PADDING
-from telegram_upload.exceptions import TelegramInvalidFile, ThumbError
-from telegram_upload.utils import scantree, truncate
-from telegram_upload.video import get_video_thumb
+from tgupi.caption_formatter import CaptionFormatter, FilePath
+from tgupi.constants import SPLIT_FILE_PART_NUMBER_PADDING
+from tgupi.exceptions import TelegramInvalidFile, ThumbError
+from tgupi.utils import scantree, truncate
+from tgupi.video import get_video_thumb
 
 mimetypes.init()
 
 
 if TYPE_CHECKING:
-    from telegram_upload._backend import TelegramManagerClient
+    from tgupi._backend import TelegramManagerClient
 
 
 def is_valid_file(file, error_logger=None):